1.完成了下拉列表的二级联动 2.完成了表格的分页 3.完成表格导出到excel

+++ src/main/java/web/action/DepartmentAction.java
@@ -1,13 +1,21 @@
 package web.action;
 
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Vector;
 
 import net.sf.json.JsonConfig;
 import net.sf.json.util.PropertyFilter;
+
+import org.apache.struts2.ServletActionContext;
+
 import pojos.Department;
 import service.DepartmentService;
+import web.ui.DepartmentUI;
+
+import common.ExcelUtil;
 
 public class DepartmentAction extends BaseAction {
 
@@ -22,6 +30,26 @@ public class DepartmentAction extends BaseAction {
 
 	public void setDepartmentService(DepartmentService departmentService) {
 		this.departmentService = departmentService;
+	}
+
+	/** 导出excel表时的输入流、文件名称;以及get和set方法  */
+	private InputStream excelStream;
+	private String fileName;
+
+	public InputStream getExcelStream() {
+		return ServletActionContext.getServletContext().getResourceAsStream("excel/"+this.fileName);
+	}
+
+	public void setExcelStream(InputStream excelStream) {
+		this.excelStream = excelStream;
+	}
+
+	public String getFileName() {
+		return fileName;
+	}
+
+	public void setFileName(String fileName) {
+		this.fileName = fileName;
 	}
 
 	/** 获取前端表单所有字段  */
@@ -123,7 +151,7 @@ public class DepartmentAction extends BaseAction {
 			this.printString(false, "获取参数错误！");
 			return null;
 		}
-		Department department =this.departmentService.find(Department.class, departmentId);
+		Department department =this.departmentService.find(departmentId);
 		department.setDepartmentName(departmentName);
 		department.setCreateTime(createTime);
 		department.setDescription(description);
@@ -148,19 +176,33 @@ public class DepartmentAction extends BaseAction {
 	 * 2014-2-11    caiwenming      v1.0.0         create
 	 */
 	public String deleteDept() {
+		
 		/**如果有多个id，则获取到的departmentIds格式是：id1,id2,id3,id4.... */
 		String[] str=this.departmentIds.split(",");
 		ArrayList<Department> departments= new ArrayList<Department>();
+		String name="";
 		/**遍历id，并实例化类型，在add到List */
 		for(int i=0;i<str.length;i++){
-			Department department = new Department();
-			/** 因为Department与其他实体类不存在多对一的关系，即数据表中没有外键，所有新建对象只许设置id即可 */
-			department.setDepartmentId(Integer.parseInt(str[i]));
+			Department department = this.departmentService.find(Integer.parseInt(str[i]));
+//			/** 因为Department与其他实体类不存在多对一的关系，即数据表中没有外键，所有新建对象只许设置id即可 */
+//			department.setDepartmentId(Integer.parseInt(str[i]));
+			if(department.getStaffs()!=null){
+				if(i==0){
+					name+=department.getDepartmentName();
+				}else {
+					name=name+"、"+department.getDepartmentName();
+				}
+			}
 			departments.add(department);
 		}
-		this.departmentService.removeAll(departments);
-		this.printString(true, "");
-		return null;
+		if (name=="") {
+			this.departmentService.removeAll(departments);
+			this.printString(true, "");
+			return null;
+		}else {
+			this.printString(false, name+"还有员工，请删除员工信息后在进行删除操作...");
+			return null;
+		}
 	}
 
 	/**
@@ -182,7 +224,6 @@ public class DepartmentAction extends BaseAction {
 		List<Department> departments = this.departmentService.findAll();
 		JsonConfig jsonConfig =new JsonConfig();
 		jsonConfig.setJsonPropertyFilter(new PropertyFilter() {
-			
 			@Override
 			public boolean apply(Object arg0, String arg1, Object arg2) {
 				if(arg1.equals("staffs")){
@@ -216,5 +257,73 @@ public class DepartmentAction extends BaseAction {
 		List<Department> departments=this.departmentService.findBysql(sql);
 		this.printList(0, 0, 0, departments);
 		return null;
+	}
+	
+	/**
+	 *
+	 * @Description:将部门数据表中记录全部或根据id导出到Excel表
+	 *
+	 * @return
+	 *
+	 * @version:v1.0
+	 * @author:caiwenming
+	 * @throws Exception 
+	 * @date:2014-2-12 上午9:58:08
+	 *
+	 * Modification History:
+	 * Date         Author        Version      Description
+	 * -----------------------------------------------------------------
+	 * 2014-2-12    caiwenming      v1.0.0         create
+	 */
+	public String exportDept() throws Exception{
+		System.out.println(departmentIds);
+		List<Department> departments= new ArrayList<Department>();
+		if(departmentIds.equals("")){
+			departments = this.departmentService.findAll();
+		}else{
+			/**如果有多个id，则获取到的departmentIds格式是：id1,id2,id3,id4.... */
+			String[] str=this.departmentIds.split(",");
+			
+			/**遍历id，并实例化类型，在add到List */
+			for(int i=0;i<str.length;i++){
+				Department department = this.departmentService.find(Integer.parseInt(str[i]));
+				departments.add(department);
+			}
+		}
+		Vector<String> head=DepartmentUI.getHead();
+		List<Vector<String>> dataList=DepartmentUI.getDataList(departments);
+		String downLoadPath =ServletActionContext.getServletContext().getRealPath("/")+"excel\\";
+		String fileName=ExcelUtil.createFileName("Department")+".xls";
+		if(ExcelUtil.printExcel(head, dataList, downLoadPath+fileName)){
+			download(fileName);
+			//System.out.println(ServletActionContext.getServletContext().getRealPath("excel/Department201402131756458884286.xls"));
+			return "success";
+		}else {
+			this.printString(false, "");
+		}
+		
+		return null;
+	}
+	
+	/**
+	 *
+	 * @Description:设置action类中的fileName和excelStream，从而使当前端点击导出时，浏览器可以识别为文件下载
+	 *
+	 * @param fileName 文件名称
+	 * @throws Exception
+	 *
+	 * @version:v1.0
+	 * @author:caiwenming
+	 * @date:2014-2-15 上午8:34:53
+	 *
+	 * Modification History:
+	 * Date         Author        Version      Description
+	 * -----------------------------------------------------------------
+	 * 2014-2-15    caiwenming      v1.0.0         create
+	 */
+	public void download(String fileName) throws Exception{
+		this.fileName=fileName;
+		this.excelStream=ServletActionContext.getServletContext().getResourceAsStream("excel/"+fileName);
+		System.out.println(this.fileName+this.excelStream);
 	}
 }
